speech-to-text-models/requirements.txt
# 핵심 의존성
openai-whisper==20240930
torch==2.5.1            
numpy==2.0.2            

# 오디오 처리
librosa==0.10.2.post1    
soundfile==0.12.1        
pydub==0.25.1           

# 데이터 처리
pandas==2.2.3           

# 유틸리티
python-dotenv==1.0.1     

# wav2vec2 관련 추가
transformers==4.37.2     
moviepy==1.0.3          
ffmpeg-python==0.2.0     